+++ subsys/net/CMakeLists.txt
@@ -1,7 +1,11 @@
 # SPDX-License-Identifier: Apache-2.0
 
 zephyr_library()
-zephyr_library_sources_ifdef(CONFIG_NET_BUF             buf.c)
+zephyr_library_sources_ifdef(
+  CONFIG_NET_BUF
+  buf.c
+  buf_simple.c
+  )
 zephyr_library_sources_ifdef(CONFIG_NET_HOSTNAME_ENABLE hostname.c)
 
 if(CONFIG_NETWORKING)
